Mode TopTabFader into its own file
